+++ BE/toDoListAPI/src/main/java/school/mindera/toDoListAPI/service/UserService.java
@@ -1,0 +1,7 @@
+package school.mindera.toDoListAPI.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class UserService {
+}
